requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
# backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
config==0.5.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
docopt==0.6.2
frozenlist==1.3.0
geographiclib==1.52
geoip2==4.5.0
geopy==2.2.0
idna==3.3
importlib-metadata==4.8.2
install==1.3.5
Js2Py==0.71
Markdown==3.3.6
maxminddb==2.2.0
multidict==6.0.2
munch==2.5.0
numpy==1.22.1
packaging==21.3
pandas==1.4.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyjsparser==2.7.1
pyparsing==3.0.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
whitenoise==6.0.0
yarl==1.7.2
zipp==3.6.0
